Add paste handler for Site Health info
- Add shared site-health-parser module with reusable parsing functions
- Update siteHealthImport step to use the shared parser
- Add Site Health detection to content-detection.ts
- Add paste handler that creates individual installPlugin/installTheme steps
  instead of a single siteHealthImport step

When pasting Site Health info, the handler parses the content and adds
individual steps for each plugin and theme found.
