+++ steps/siteHealthImport.ts
@@ -1,5 +1,14 @@
 import type { StepFunction, BlueprintStep, StepResult } from './types.js';
 import type { BlueprintV2Declaration } from '@wp-playground/blueprints';
+import {
+	parseSiteHealth,
+	parseSections,
+	parsePluginsFromLines,
+	parseThemeFromLines,
+	deriveThemeSlug,
+	deriveSlugFromName,
+	type PluginInfo,
+} from '../src/shared/site-health-parser.js';
 
 export interface SiteHealthImportStep extends BlueprintStep {
 	vars?: {
@@ -18,199 +27,6 @@ interface ParsedSiteHealthData {
 	options: Record<string, string>;
 }
 
-interface PluginInfo {
-	name: string;
-	version: string;
-	author: string | null;
-}
-
-interface ThemeInfo {
-	name: string | null;
-	version: string | null;
-	path: string | null;
-}
-
-/**
- * Derive a WordPress slug from a plugin or theme name.
- */
-function deriveSlugFromName(name: string): string {
-	return name
-		.split(':')[0]
-		.trim()
-		.toLowerCase()
-		.replace(/\s+/g, '-')
-		.replace(/[^a-z0-9-]/g, '')
-		.replace(/-+/g, '-')
-		.replace(/^-|-$/g, '');
-}
-
-/**
- * Clean site health content by stripping leading backticks.
- */
-function cleanSiteHealthContent(content: string): string {
-	const trimmed = content.trim();
-	return trimmed.startsWith('`') ? trimmed.slice(1).trim() : trimmed;
-}
-
-/**
- * Parse site health content into sections.
- * Returns a map of section name to array of lines.
- */
-function parseSections(content: string): Map<string, string[]> {
-	const sections = new Map<string, string[]>();
-	const cleanContent = cleanSiteHealthContent(content);
-	const parts = cleanContent.split(/^###\s+/m);
-
-	for (const part of parts) {
-		if (!part.trim()) continue;
-
-		const lines = part.split('\n');
-		const sectionNameRaw = lines[0].trim().replace(/\s*###\s*$/, '');
-		const sectionName = sectionNameRaw.split(' ')[0];
-
-		sections.set(sectionName, lines.slice(1));
-	}
-
-	return sections;
-}
-
-/**
- * Parse theme info from wp-active-theme section lines.
- */
-function parseThemeFromLines(lines: string[]): ThemeInfo {
-	const info: ThemeInfo = { name: null, version: null, path: null };
-
-	for (const line of lines) {
-		const match = line.match(/^(\w+):\s*(.+)$/);
-		if (!match) continue;
-
-		const key = match[1];
-		const value = match[2].trim();
-
-		if (key === 'name') {
-			const slugMatch = value.match(/\(([^)]+)\)$/);
-			info.name = slugMatch ? slugMatch[1] : value;
-		} else if (key === 'version') {
-			info.version = value;
-		} else if (key === 'theme_path') {
-			const pathMatch = value.match(/\/([^/]+)\/?$/);
-			if (pathMatch) info.path = pathMatch[1];
-		}
-	}
-
-	return info;
-}
-
-/**
- * Derive theme slug from theme info.
- */
-function deriveThemeSlug(info: ThemeInfo): string | null {
-	if (info.name && info.name === info.name.toLowerCase() && !info.name.includes(' ')) {
-		return info.name;
-	}
-	if (info.path) {
-		return info.path;
-	}
-	if (info.name) {
-		return info.name.toLowerCase().replace(/\s+/g, '-');
-	}
-	return null;
-}
-
-/**
- * Parse a plugin info from a wp-plugins-active line.
- */
-function parsePluginFromLine(line: string): PluginInfo | null {
-	if (!line.trim()) return null;
-
-	const versionIndex = line.indexOf('version:');
-	if (versionIndex === -1) return null;
-
-	const pluginNamePart = line.substring(0, versionIndex).trim();
-	const pluginName = pluginNamePart.replace(/\s*:\s*$/, '');
-
-	const versionPart = line.substring(versionIndex + 'version:'.length).trim();
-	const versionMatch = versionPart.match(/^([^,]+)/);
-	if (!versionMatch) return null;
-
-	const pluginVersion = versionMatch[1].trim();
-
-	let pluginAuthor: string | null = null;
-	const authorIndex = line.indexOf('author:');
-	if (authorIndex !== -1) {
-		const authorPart = line.substring(authorIndex + 'author:'.length).trim();
-		const updatesMatch = authorPart.match(/^(.+?)(?:,\s*(?:Updates|Auto-updates))/);
-		pluginAuthor = updatesMatch ? updatesMatch[1].trim() : authorPart.trim();
-	}
-
-	return { name: pluginName, version: pluginVersion, author: pluginAuthor };
-}
-
-/**
- * Parse plugins from wp-plugins-active section lines.
- */
-function parsePluginsFromLines(lines: string[]): PluginInfo[] {
-	const plugins: PluginInfo[] = [];
-	for (const line of lines) {
-		const plugin = parsePluginFromLine(line);
-		if (plugin) plugins.push(plugin);
-	}
-	return plugins;
-}
-
-interface SyncParsedData {
-	theme: string | null;
-	plugins: string[];
-}
-
-/**
- * Synchronous parsing of site health content for theme and plugin slugs.
- * Used by toV1() and toV2() which cannot be async.
- */
-function parseSiteHealthSync(
-	content: string,
-	options: { installTheme: boolean; installPlugins: boolean; installLatest: boolean }
-): SyncParsedData {
-	const result: SyncParsedData = { theme: null, plugins: [] };
-
-	if (!content.trim()) {
-		return result;
-	}
-
-	const sections = parseSections(content);
-
-	if (options.installTheme) {
-		const themeLines = sections.get('wp-active-theme');
-		if (themeLines) {
-			const themeInfo = parseThemeFromLines(themeLines);
-			const themeSlug = deriveThemeSlug(themeInfo);
-			if (themeSlug) {
-				result.theme = options.installLatest || !themeInfo.version
-					? themeSlug
-					: `${themeSlug}@${themeInfo.version}`;
-			}
-		}
-	}
-
-	if (options.installPlugins) {
-		const pluginLines = sections.get('wp-plugins-active');
-		if (pluginLines) {
-			const pluginInfos = parsePluginsFromLines(pluginLines);
-			for (const plugin of pluginInfos) {
-				const slug = deriveSlugFromName(plugin.name);
-				if (slug) {
-					const pluginSpec = options.installLatest || !plugin.version
-						? slug
-						: `${slug}@${plugin.version}`;
-					result.plugins.push(pluginSpec);
-				}
-			}
-		}
-	}
-
-	return result;
-}
-
 /**
  * Query WordPress.org plugins update-check API to map plugin names to slugs.
  * Uses the update-check API which accepts plugin metadata (Name, Author, Version)
@@ -224,7 +40,6 @@ async function mapPluginNamesToSlugs(plugins: PluginInfo[]): Promise<Map<string,
 	}
 
 	try {
-		// Build request body for update-check API
 		const pluginsData: Record<string, { Name: string; Version: string; Author?: string }> = {};
 		const activePlugins: string[] = [];
 
@@ -276,8 +91,6 @@ async function mapPluginNamesToSlugs(plugins: PluginInfo[]): Promise<Map<string,
 		console.warn('Warning: Could not query WordPress.org plugins update-check API:', error);
 	}
 
-	// For plugins not mapped by the update-check API, derive slugs optimistically
-	// and verify they exist on WordPress.org before including them
 	const unmappedPlugins = plugins.filter((plugin) => !nameToSlug.has(plugin.name));
 
 	if (unmappedPlugins.length > 0) {
@@ -285,7 +98,6 @@ async function mapPluginNamesToSlugs(plugins: PluginInfo[]): Promise<Map<string,
 			.map((plugin) => ({ plugin, slug: deriveSlugFromName(plugin.name) }))
 			.filter(({ slug }) => slug);
 
-		// Verify each candidate slug exists on WordPress.org
 		const verificationPromises = candidateSlugs.map(async ({ plugin, slug }) => {
 			try {
 				const response = await fetch(
@@ -299,7 +111,6 @@ async function mapPluginNamesToSlugs(plugins: PluginInfo[]): Promise<Map<string,
 
 				if (response.ok) {
 					const data = await response.json();
-					// The API returns an object with 'error' if the plugin doesn't exist
 					if (data && !data.error && data.slug) {
 						return { plugin, slug: data.slug, verified: true };
 					}
@@ -339,7 +150,6 @@ async function parseSiteHealthContent(content: string): Promise<ParsedSiteHealth
 
 	const sections = parseSections(content);
 
-	// Parse wp-core section
 	const coreLines = sections.get('wp-core');
 	if (coreLines) {
 		for (const line of coreLines) {
@@ -361,7 +171,6 @@ async function parseSiteHealthContent(content: string): Promise<ParsedSiteHealth
 		}
 	}
 
-	// Parse wp-server section
 	const serverLines = sections.get('wp-server');
 	if (serverLines) {
 		for (const line of serverLines) {
@@ -376,7 +185,6 @@ async function parseSiteHealthContent(content: string): Promise<ParsedSiteHealth
 		}
 	}
 
-	// Parse wp-active-theme section
 	const themeLines = sections.get('wp-active-theme');
 	if (themeLines) {
 		const themeInfo = parseThemeFromLines(themeLines);
@@ -386,7 +194,6 @@ async function parseSiteHealthContent(content: string): Promise<ParsedSiteHealth
 		}
 	}
 
-	// Parse wp-plugins-active section
 	const pluginLines = sections.get('wp-plugins-active');
 	if (pluginLines) {
 		const pluginInfos = parsePluginsFromLines(pluginLines);
@@ -415,7 +222,7 @@ export const siteHealthImport: StepFunction<SiteHealthImportStep> = (step: SiteH
 
 	return {
 		toV1() {
-			const { theme, plugins } = parseSiteHealthSync(siteHealthContent, {
+			const { theme, plugins } = parseSiteHealth(siteHealthContent, {
 				installTheme,
 				installPlugins,
 				installLatest,
@@ -459,7 +266,7 @@ export const siteHealthImport: StepFunction<SiteHealthImportStep> = (step: SiteH
 		},
 
 		toV2(): BlueprintV2Declaration {
-			const { theme, plugins } = parseSiteHealthSync(siteHealthContent, {
+			const { theme, plugins } = parseSiteHealth(siteHealthContent, {
 				installTheme,
 				installPlugins,
 				installLatest,
@@ -509,5 +316,4 @@ siteHealthImport.vars = [
 	},
 ];
 
-// Export the async parsing function for potential use in other contexts
 export { parseSiteHealthContent, mapPluginNamesToSlugs };
